feat: Introduce new combat character sheet components and game services, refactor shared schemas, and update agent rules.

+++ backend/src/lifecycle/graphql/type-defs.ts
@@ -44,7 +44,7 @@ export const typeDefs = `
     seed: String
     chunkSize: Int
     globalScale: Float
-    seaLevel: Int
+    seaLevel: Float
     elevationScale: Float
     roughness: Float
     detail: Float
